// src/components/weather-details/types.d.ts
/* Type for current weather details  */
export type CurrentWeatherDetails = {
  date: string;
  temprature: string;
  feelsLike: string;
  humidity: string;
  dew_point: string;
  visibility: string;
  lat: string;
  long: string;
  country: string;
  name: string;
  weatherCondition: string;
  windSpeed: string;
};